better appearance for createRecipe and autpmatic deletion for recipePicture that become unnecessary
