+++ foodapp/src/App.js
@@ -8,7 +8,7 @@ import { Container } from '@mui/material'
 import RegistrationForm from './pages/RegistrationPage'
 import UsersPage from './pages/UsersPage'
 import UserProfile from './pages/UserProfilePage'
-import CreateRecipePage from './pages/createRecipe'
+import CreateRecipePage from './pages/CreateRecipe'
 import UserSearchPage from './pages/UserRecipeSearch'
 import PersistedLogin from './components/PersistedLogin'
 import ErrorLayout from './components/ErrorLayout'
